Fixed (probably) writing scores to file

+++ core/src/com/gdx/tetris/GameOverScreen.java
@@ -96,53 +96,60 @@ public class GameOverScreen implements Screen {
 	    });
 	    table.add(blankField).height(20).row();
 	    table.add(menu).width(300).height(100);
-	    
+	    //Writing score
 	    try {
 	        File scores = new File("scores.txt"), 
 	        	 scores2 = new File("temp_scores.txt");
 	        //Create and open new temporary score file
 	        if (scores2.createNewFile()) {
-	            System.out.println("File created: " + scores.getName());
+	            System.out.println("Scores2: File created: " + scores2.getName());
 	          } else {
-	            System.out.println("File already exists.");
+	            System.out.println("Scores2: File already exists.");
 	          }
-	        
-	        //Read data from prev score file
-	        Scanner myReader = new Scanner(scores);
-	        FileWriter myWriter = new FileWriter(scores2.getAbsolutePath());
-	        DateTimeFormatter dtf = DateTimeFormatter.ofPattern("yyyy/MM/dd HH:mm:ss");  
-	        LocalDateTime now = LocalDateTime.now(); 
-	        int n = 0;
-	        boolean inserted = false;
-	        int [] sc = new int [50];
-            while (myReader.hasNextLine() && n != 50) {
-              String data = myReader.nextLine();
-              //Starting from 23rd pos, read number to an array
-              int thisScore = 0;
-              for(int i = 29; i < data.length(); i++) {
-            	  thisScore*=10;
-            	  thisScore+=Character.getNumericValue(data.charAt(i));
-              }
-              sc[n] = thisScore;
-              //------------------
-              if(thisScore <= score && !inserted) {
-            	  myWriter.write(dtf.format(now) + "   Score: " + score + "\n");
-            	  inserted = true;
-              }
-              myWriter.write(data + '\n');
-              System.out.println(data);
-              n++;
-            }
-            myReader.close();
-	        
+	       
+	        //Read data from prev score file                                              
+	        Scanner myReader = new Scanner(scores);                                       
+	        FileWriter myWriter = new FileWriter(scores2.getAbsolutePath());              
+	        DateTimeFormatter dtf = DateTimeFormatter.ofPattern("yyyy/MM/dd HH:mm:ss");   
+	        LocalDateTime now = LocalDateTime.now();                                      
+	        if(!scores.exists()) {
+		        int n = 0;
+		        boolean inserted = false;
+		        int [] sc = new int [50];
+	            while (myReader.hasNextLine() && n != 50) {
+	              String data = myReader.nextLine();
+	              //Starting from 23rd pos, read number to an array
+	              int thisScore = 0;
+	              for(int i = 29; i < data.length(); i++) {
+	            	  thisScore*=10;
+	            	  thisScore+=Character.getNumericValue(data.charAt(i));
+	              }
+	              sc[n] = thisScore;
+	              //------------------
+	              if(thisScore <= score && !inserted) {
+	            	  myWriter.write(dtf.format(now) + "   Score: " + score + "\n");
+	            	  inserted = true;
+	              }
+	              myWriter.write("Scores: " + data + '\n');
+	              System.out.println(data);
+	              n++;
+	            }
+	            myReader.close();
+		        
+	        }
+	        else {
+	        	myWriter.write(dtf.format(now) + "   Score: " + score + "\n");
+	        }
             //Re-create it
-	        scores.delete();
-	        scores2.renameTo(scores);
-	        
 	        myWriter.close();
-	        
+	        System.out.println("Deleting scores.txt...\n");
+	        if(scores.exists()) scores.delete();
+	        System.out.println("Renaming temp_scores.txt -> scores.txt...\n");
+        	scores2.renameTo(scores);
+        	 System.out.println("Done.\n");
+	         
 	    } catch (IOException e) {
-	          System.out.println("An error occurred.");
+	          System.out.println("Scores: An error occurred.");
 	          e.printStackTrace();
 	    }
